+++ src/component/bindings.js
@@ -1,0 +1,13 @@
+export default {
+  startDate: '@',
+  endDate: '@',
+  title: '@',
+  description: '@',
+  location: '@',
+  btnText: '@',
+  hoverText: '@',
+  format: '@',
+  className: '@',
+  useCaret: '@',
+  useBootstrap: '@'
+}