+++ webpack.config.js
@@ -1,0 +1,164 @@
+'use strict';
+
+// Modules
+var webpack = require('webpack');
+var autoprefixer = require('autoprefixer');
+var HtmlWebpackPlugin = require('html-webpack-plugin');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
+var VendorChunkPlugin = require('webpack-vendor-chunk-plugin');
+var CopyWebpackPlugin = require('copy-webpack-plugin');
+
+/**
+ * Get dependencies list from package.json
+ */
+var vendorDeps = require('./package.json').dependencies;
+
+/**
+ * Env
+ * Get npm lifecycle event to identify the environment
+ */
+var ENV = process.env.npm_lifecycle_event;
+var isTest = ENV === 'test' || ENV === 'test-watch';
+var isProd = ENV === 'build';
+
+module.exports = function makeWebpackConfig() {
+  /**
+   * Config
+   * Reference: http://webpack.github.io/docs/configuration.html
+   * This is the object where all configuration gets set
+   */
+  var config = {};
+
+  /**
+   * Entry
+   * Reference: http://webpack.github.io/docs/configuration.html#entry
+   * Should be an empty object if it's generating a test build
+   * Karma will set this when it's a test build
+   */
+  config.entry = isTest ? void 0 : {
+    addtocalendar: './src/component/index.js',
+    vendor: Object.keys(vendorDeps)
+  };
+
+  /**
+   * Output
+   * Reference: http://webpack.github.io/docs/configuration.html#output
+   * Should be an empty object if it's generating a test build
+   * Karma will handle setting it up for you when it's a test build
+   */
+  config.output = isTest ? {} : {
+    // Absolute output directory
+    path: __dirname + '/dist',
+
+    // Output path from the view of the page
+    // Uses webpack-dev-server in development
+    publicPath: isProd ? '/' : 'http://localhost:8080/',
+
+    // Filename for entry points
+    // Only adds hash in build mode
+    filename: '[name].js',
+
+    // Filename for non-entry points
+    // Only adds hash in build mode
+    chunkFilename: '[name].js'
+  };
+
+  /**
+   * Loaders
+   * Reference: http://webpack.github.io/docs/configuration.html#module-loaders
+   * List: http://webpack.github.io/docs/list-of-loaders.html
+   * This handles most of the magic responsible for converting modules
+   */
+
+  // Initialize module
+  config.module = {
+    rules: [{
+      // JS LOADER
+      // Reference: https://github.com/babel/babel-loader
+      // Transpile .js files using babel-loader
+      // Compiles ES6 and ES7 into ES5 code
+      test: /\.js$/,
+      loader: 'babel-loader',
+      exclude: /node_modules/
+    }, {
+      test: /\.scss$/,
+
+      loader: isTest ? 'null-loader' : ExtractTextPlugin.extract({
+        fallback: 'style-loader',
+        use: [
+          'css-loader',
+          'sass-loader'
+        ]
+      })
+    }, {
+      test: /\.html$/,
+      loader: 'raw-loader'
+    }]
+  };
+
+  // ISTANBUL LOADER
+  // https://github.com/deepsweet/istanbul-instrumenter-loader
+  // Instrument JS files with istanbul-lib-instrument for subsequent code coverage reporting
+  // Skips node_modules and files that end with .spec.js
+  if (isTest) {
+    config.module.rules.push({
+      enforce: 'pre',
+      test: /\.js$/,
+      exclude: [/node_modules/],
+      loader: 'istanbul-instrumenter-loader',
+      query: {
+        esModules: true
+      }
+    })
+  }
+
+  config.plugins = [];
+
+  // Skip rendering index.html in test mode
+  if (!isTest) {
+    // Reference: https://github.com/ampedandwired/html-webpack-plugin
+    // Render index.html
+    config.plugins.push(
+      new HtmlWebpackPlugin({
+        template: './src/public/index.html',
+        inject: 'body'
+      }),
+
+      new ExtractTextPlugin({
+        filename: '[name].css',
+        disable: !isProd,
+        allChunks: true
+      })
+    )
+  }
+
+  // Add build specific plugins
+  if (isProd) {
+    config.plugins.push(
+      // Reference: http://webpack.github.io/docs/list-of-plugins.html#noerrorsplugin
+      // Only emit files when there are no errors
+      new webpack.NoEmitOnErrorsPlugin(),
+
+      new webpack.optimize.UglifyJsPlugin(),
+      // copy sass file to dist
+      new CopyWebpackPlugin([{
+        from: __dirname + '/src/component/styles.scss',
+        to: __dirname + '/dist/addtocalendar.scss'
+      }]),
+      new webpack.optimize.CommonsChunkPlugin('vendor', 'vendor.js'),
+      new VendorChunkPlugin('vendor')
+    )
+  }
+
+  /**
+   * Dev server configuration
+   * Reference: http://webpack.github.io/docs/configuration.html#devserver
+   * Reference: http://webpack.github.io/docs/webpack-dev-server.html
+   */
+  config.devServer = {
+    contentBase: './src/public',
+    stats: 'minimal'
+  };
+
+  return config;
+}();