+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.3.0
+
+- removes Bootstrap as a dependency
+- implements custom styles and style extensibility
+- implements sass, webpack, and other build tools
+- enhances tests
+- fixes broken demo (#36)
+- adds missing PRODID, UID, and DTSTAMP fields for proper ics validation (#24)
+
 ## v1.2.2
 
 - updates docs